rendering: refactor console out of overlay
